requirements.txt
anyio==3.6.1
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
chardet==5.0.0
click==8.1.3
cycler==0.11.0
essentia==2.1b6.dev858
fastapi==0.84.0
Flask==2.1.2
fonttools==4.33.3
h11==0.13.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
more-itertools==8.13.0
music21==5.5.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
webcolors==1.12
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
